+++ src/main/java/org/example/dto/TransactionIdsResponse.java
@@ -2,5 +2,5 @@ package org.example.dto;
 
 import java.util.List;
 
-public record TransactionIdsResponse(List<Integer> transactionIds) {
+public record TransactionIdsResponse(List<Long> transactionIds) {
 }
