+++ src/main/java/org/example/dto/TransactionRequest.java
@@ -1,3 +1,3 @@
 package org.example.dto;
 
-public record TransactionRequest(int amount, String type, Integer parentId){}
+public record TransactionRequest(double amount, String type, Long parentId){}
